More tests to show that the link regex is too eager.
