+++ tests/html-fixtures/figure.rst
