+++ tests/html-fixtures/simple-table.rst
